Moved documents uploaded to PaymentReview dir
