+++ src/scenes/Moves/Ppm/ExpensesLanding.jsx
@@ -9,7 +9,7 @@ import WizardHeader from '../WizardHeader';
 
 import './Expenses.css';
 import { connect } from 'react-redux';
-import DocumentsUploaded from './DocumentsUploaded';
+import DocumentsUploaded from './PaymentReview/DocumentsUploaded';
 
 const reviewPagePath = '/ppm-payment-review';
 const nextPagePath = '/ppm-expenses';
@@ -49,7 +49,9 @@ class ExpensesLanding extends Component {
             </ProgressTimeline>
           }
         />
-        <DocumentsUploaded moveId={moveId} />
+        <div className="usa-grid">
+          <DocumentsUploaded showToggleDocs moveId={moveId} />
+        </div>
 
         <div className="usa-grid expenses-container">
           <h3 className="expenses-header">Do you have any storage or moving expenses?</h3>
